Fix - dto 수정(orderItemQuantity 컬럼을 Integer 타입으로 변경)

+++ src/main/java/com/deliveryapp/catchabite/dto/OrderItemDTO.java
@@ -36,6 +36,6 @@ public class OrderItemDTO {
     // 수량은 보통 0이상 또는 일반적으로 1이상
     @NotNull
     @Positive
-    private Long orderItemQuantity;
+    private Integer orderItemQuantity;
     
 }
